Fix startup of gmediarender
gmediarender doesn't like logging to stdout by open()-ing /dev/stdout. This
works on the command line but not when run under systemd.  I forked it and
added a patch to accept a log file of '-' to mean 'use stdout'.

This change means that we can avoid forking in systemd.